mkfs: write a valid root directory
Writing a valid root directory inode in the second block of the disk. For
now with same default parameters as superblock and no data.

This will fail to be read, since there is no data contained in the inode.

ls: reading directory /mnt/befs/: Input/output error
total 0

The good news is that now the befs file system mounts.