mkfs: get block_shift from block_size
Generate the block_shift for the block_size. For now the block_size is
hardcoded but that is just temporary, so it is a good idea to have the
block shift dynamically generated.

Instead of using ffs(), since we know the list of potential block sizes.
I could've written the simple loop below that expects 2^10, 2^11, or 2^12.
But decided to keep it more readable by just using ffs.

for (x = 12; x > 9; x--) {
    if (BLOCK_SIZE >> x) {
        block_shift = x;
        break;
    }
}

Kudos to Derek Foreman for letting me know about ffs()