mkfs: move root dir to 0x805
The root directory for all BFS file systems begins in the eighth allocation
group (i.e., starting at block 65,536)

Since BeFS reserves the first eight allocation groups as the preferred area for
allocation indices and their data. BFS reserves these eight allocation groups
simply by convention.

All data blocks for a directory are allocated from the same allocation group
as the directory i-node (if possible). File i-nodes are also put in the same
allocation group as the directory that contains them. The result is that
directory data and i-node blocks for the files in the directory will be near
each other.